backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
groq>=0.4.0
ultralytics>=8.0.173
opencv-python-headless>=4.8.0
tensorflow>=2.13.0
web3>=6.0.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
